Added Messages and confirmation blocks. Fixed bugs 1.Sound playing while app in background. 2. Sucess Label missing if user deletes record. 3.when user uploads image - crop image is directed to screen with only cross and right icon.It should display a message so taht user understands what operation he has to perform. 4.when user uploads image the recently uploaded image should be displayed on the main frame screen where all the images are being displayed. 5. Image counter should be there so taht user should know how many images he has uploaded. 6.Screen flashes if user clicks on back button or settings icon. 7.label should appear is user mutes or deletes record. 8.if user sitch languga eto germany and click on camera icon ,cancel text appears in english on popup

+++ YourSPBall/YourSPBall/Views/EditBallDrawingPage.xaml.cs
@@ -1,6 +1,7 @@
 ﻿using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
 using YourSPBall.Models;
+using YourSPBall.Resources;
 
 namespace YourSPBall
 {
@@ -100,10 +101,16 @@ namespace YourSPBall
                 return new Command(() =>
                 {
                     App.IconClicked();
-                    
+
+                    string action = DisplayActionSheet(AppResources.DeleteBallMsg, AppResources.Cancel, null, new string[] { AppResources.No, AppResources.Yes }).Result;
+
+                    if (action != AppResources.Yes)
+                        return;
+
                     if (SPBall.ID > 0)
                         App.Database.DeleteSPBall(SPBall);
 
+                    DisplayAlert("YourSPBall", AppResources.DeleteDataMsg, AppResources.Cancel);
                     App.Current.MainPage = new NavigationPage(new MainMenuPage());
                 });
             }
